Fixed a heading issue.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@5094 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ trajectory_rollout/include/trajectory_rollout/governor_node.h
@@ -105,7 +105,7 @@ class WavefrontMapAccessor : public costmap_2d::ObstacleMapAccessor {
       resolution_ = res;
     }
 
-    void updateSize(double width, double height){
+    void updateSize(unsigned int width, unsigned int height){
       width_ = width;
       height_ = height;
     }
